requirements.txt
selenium
webdriver-manager
fpdf
Pillow
pytest
pytest-html